feeds_service/requirements.txt
asgiref==3.6.0
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==3.2
django-3-jet==1.0.8
django-material-admin==1.8.6
falcon==3.1.1
falcon-swagger-ui==1.2.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jsonschema==2.5.1
MarkupSafe==2.1.2
msgpack-python==0.4.8
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-mimeparse==1.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.1.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
mysqlclient==2.1.1
